refactor: split monolithic routes file into domain-specific files (ENG-261)
Split the 604-line guards/index.tsx into focused route config files:
- auth-routes.tsx: sign-in, sign-up, password reset, email verification
- platform-routes.tsx: all /platform/* admin routes
- project-routes.tsx: project dashboard, builder, and settings routes
- public-routes.tsx: embed, templates, forms, chats, redirect, 404

guards/index.tsx now only composes route arrays and exports the router (~50 lines).
