+++ packages/web/src/app/routes/auth-routes.tsx
@@ -1,0 +1,59 @@
+import { PageTitle } from '@/app/components/page-title';
+import { VerifyEmail } from '@/features/authentication/components/verify-email';
+import { AcceptInvitation } from '@/features/members/component/accept-invitation';
+
+import { ChangePasswordPage } from './change-password';
+import { ResetPasswordPage } from './forget-password';
+import { SignInPage } from './sign-in';
+import { SignUpPage } from './sign-up';
+
+export const authRoutes = [
+  {
+    path: '/forget-password',
+    element: (
+      <PageTitle title="Forget Password">
+        <ResetPasswordPage />
+      </PageTitle>
+    ),
+  },
+  {
+    path: '/reset-password',
+    element: (
+      <PageTitle title="Reset Password">
+        <ChangePasswordPage />
+      </PageTitle>
+    ),
+  },
+  {
+    path: '/sign-in',
+    element: (
+      <PageTitle title="Sign In">
+        <SignInPage />
+      </PageTitle>
+    ),
+  },
+  {
+    path: '/verify-email',
+    element: (
+      <PageTitle title="Verify Email">
+        <VerifyEmail />
+      </PageTitle>
+    ),
+  },
+  {
+    path: '/sign-up',
+    element: (
+      <PageTitle title="Sign Up">
+        <SignUpPage />
+      </PageTitle>
+    ),
+  },
+  {
+    path: '/invitation',
+    element: (
+      <PageTitle title="Accept Invitation">
+        <AcceptInvitation />
+      </PageTitle>
+    ),
+  },
+];
